Add feature that checks user id in editing and removing reviews

+++ app.py
@@ -1,6 +1,6 @@
 import sqlite3
 from flask import Flask
-from flask import redirect, render_template, request, session
+from flask import abort, redirect, render_template, request, session
 from werkzeug.security import check_password_hash, generate_password_hash
 import config
 import db
@@ -42,11 +42,16 @@ def create_item():
 @app.route("/edit_item/<int:item_id>")
 def edit_item(item_id):
     item = reviews.get_review(item_id)
+    if item["user_id"] != session["user_id"]:
+        abort(403)
     return render_template("edit_item.html", item=item)
 
 @app.route("/update_item", methods=["POST"]) 
 def update_item():
     item_id = request.form["item_id"]
+    item = reviews.get_review(item_id)
+    if item["user_id"] != session["user_id"]:
+        abort(403)
     title = request.form["title"]
     review_text = request.form["review_text"]
 
@@ -56,9 +61,14 @@ def update_item():
 
 @app.route("/delete_item/<int:item_id>", methods=["GET", "POST"])
 def delete_item(item_id):
+
+    item = reviews.get_review(item_id)
+    if item["user_id"] != session["user_id"]:
+        abort(403)
+
     if request.method == "GET":
-        item = reviews.get_review(item_id)
         return render_template("delete_item.html", item=item)
+
     if request.method == "POST":
         if "delete" in request.form:
             reviews.delete_review(item_id)
